Unsplit the controller
