+++ routes/base.php
@@ -9,9 +9,6 @@
 Route::get('/', 'IndexController@getIndex')->name('index');
 Route::get('/status/{server}', 'IndexController@status')->name('index.status');
 
-// Used when linking account to oauth2
-Route::get('/oauth2/callback', 'OAuth2Controller@login')->name('oauth2.callback');
-
 /*
 |--------------------------------------------------------------------------
 | Account Controller Routes
